+++ doc/whatsnew.rst
@@ -15,6 +15,9 @@ v0.8.dev
 
 - Enhance :func:`pyriemann.utils.mean.mean_covariance` to support "power" and "poweuclid" metrics. :pr:`329` by :user:`qbarthelemy`
 
+- Add :func:`pyriemann.utils.tangentspace.exp_map_wasserstein`, :func:`pyriemann.utils.tangentspace.log_map_wasserstein`.
+  :pr:`322` by :user:`gabelstein`
+
 v0.7 (October 2024)
 -------------------